requirements.txt
flask-security
flask-sqlalchemy
flask-script
flask-bootstrap
flask-migrate
Flask-Failsafe
flask-admin
mysqlclient
bcrypt
